pi-companion/requirements.txt
python-socketio[client]>=5.0
python-dotenv>=1.0
rpi_ws281x>=5.0
